Not quite sure about the game message layout
Basically all other flags than MSGFLAG_VITAL are boring
but then flag MSGFLAG_FLUSH tells you wether
this message is usually seen as a single message in a packet
or combined with others which might be interesting.

And the MSGFLAG_NORECORD, MSGFLAG_RECORD or MSGFLAG_NOSEND
I did not even research yet but highly assume they are not
affecting the network traffic.

So I also thought about instead of saying:

+--------+---------------+
| Flags: | MSGFLAG_VITAL |
+--------+---------------+

Saying:

+--------+------+
| Vital: | true |
+--------+------+

To make it a bit easier on the eye and make it
more obvious when the flag vital is missing.
But then this is so hard to search using
ctrl+f it doesnt match.
So if one wants to check which messages
affect the sequence number it becomes annoying
to check.